requirements.txt
python-telegram-bot==20.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
asyncpg==0.29.0
aiohttp==3.9.1
